+++ CodingDojo/HarryPotter/BuyingSet.cs
@@ -6,13 +6,13 @@ namespace HarryPotter
     public class BuyingSet
     {
 
-        public double Discount { get; set; }
+        public decimal Discount { get; set; }
         public List<BuyingItem> Items { get; set; }
 
         public BuyingSet()
         {
             Items = new List<BuyingItem>();
-            Discount = 0.0;
+            Discount = 0.0M;
         }
 
         public BuyingSet CheckValid(List<BuyingItem> currentBuying)
